+++ src/components/bar/bar.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-nested-ternary, prefer-arrow-callback, no-underscore-dangle */
 // Other Shared Imports
 import { Environment as env } from '../../utils/environment';
 import * as debug from '../../utils/debug';
@@ -190,7 +191,7 @@ Bar.prototype = {
     }
 
     // Get the Legend Series
-    const series = dataset.map(d => ({ name: d.name, color: d.color, pattern: d.pattern }));
+    const series = dataset.map(d => ({ name: d.name, color: d.color, pattern: d.pattern, attributes: d.attributes }));
 
     // Ellipsis
     this.ellipsis = { use: false, percentageWidth: 25, threshold: 12 };
@@ -417,6 +418,14 @@ Bar.prototype = {
       })
       .enter()
       .append('rect')
+      .call((d) => {
+        d._groups.forEach((bars) => {
+          bars.forEach((bar) => {
+            const dat = bar.__data__;
+            utils.addAttributes($(bar), dat, dat.attributes, 'bar');
+          });
+        });
+      })
       .attr('class', (d, i) => `bar series-${i}`)
       .style('fill', (d, i) => (s.isStacked ? // eslint-disable-line
         (series.length === 1 ? (charts.chartColor(i, 'bar-single', d)) :  // eslint-disable-line
@@ -719,7 +728,14 @@ Bar.prototype = {
 
     // Add Legends
     if (s.showLegend) {
-      charts.addLegend((s.isStacked ? series : legendMap), s.type, s, this.element);
+      let legendSeries = s.isStacked ? series : legendMap;
+      legendSeries = legendSeries.map((d) => {
+        if (d.attributes && !d.data?.attributes) {
+          d.data = $.extend({}, (d.data || {}), { attributes: d.attributes });
+        }
+        return d;
+      });
+      charts.addLegend(legendSeries, s.type, s, this.element);
     }
     charts.appendTooltip();